+++ sickrage/core/search.py
@@ -167,15 +167,15 @@ def pickBestResult(results, show):
             continue
 
         # check if seeders meet out minimum requirements, disgard result if it does not
-        if hasattr(cur_result.provider, 'minseed') and cur_result.seeders not in (-1, None, ''):
-            if int(cur_result.seeders) < int(cur_result.provider.minseed):
+        if hasattr(cur_result.provider, 'minseed') and cur_result.seeders not in (-1, None):
+            if int(cur_result.seeders) < min(cur_result.provider.minseed, 1):
                 sickrage.app.log.info("Discarding torrent because it doesn't meet the minimum seeders: {}. Seeders:  "
                                       "{}".format(cur_result.name, cur_result.seeders))
                 continue
 
         # check if leechers meet out minimum requirements, disgard result if it does not
-        if hasattr(cur_result.provider, 'minleech') and cur_result.leechers not in (-1, None, ''):
-            if int(cur_result.leechers) < int(cur_result.provider.minleech):
+        if hasattr(cur_result.provider, 'minleech') and cur_result.leechers not in (-1, None):
+            if int(cur_result.leechers) < min(cur_result.provider.minleech, 0):
                 sickrage.app.log.info("Discarding torrent because it doesn't meet the minimum leechers: {}. Leechers:  "
                                       "{}".format(cur_result.name, cur_result.leechers))
                 continue